feat: Modularizando endpoint websocket
